record equality in prelude
