+++ src/boot/s/boot.min
@@ -35,15 +35,11 @@
 (define-values (generics-accessor) (lambda (generic name) (if ($generics-type? generic) (void) (let-values () (error (quote generics-accessor) "expected a generics interface" generic))) (if (symbol? name) (void) (let-values () (error (quote generics-accessor) "expected a generics interface" generic))) (letrec-values (((idx) (index-of name ($generics-type-method-names generic))) (() (let-values () (if idx (void) (let-values () (error (quote generics-accessor) "method name not associated with generics interface" name))) (values)))) (lambda (rtd) (if (record-type-descriptor? rtd) (void) (let-values () (error (quote generics-accessor) "expected a record type descriptor" rtd))) (letrec-values (((rtd-generics) (hashtable-ref $record-generics rtd (quote ()))) ((maybe-entry) (assoc generic rtd-generics))) (if maybe-entry (list-ref (cdr maybe-entry) idx) (error (quote generics-accessor) "generic not associated with record type" ($generics-type-name generic) rtd)))))))
 (define-values (generic-instance?) (lambda (generic) (if ($generics-type? generic) (void) (let-values () (error (quote generic-instance?) "expected a generics interface" generic))) (lambda (rtd) (if (record-type-descriptor? rtd) (let-values (((generics) (hashtable-ref $record-generics rtd #f))) (if generics (assoc generic generics) #f)) #f))))
 (define-values (gen:equal+hash gen:equal+hash?) (let-values (((:gen) ($make-generics-type (quote equal+hash) (quote ((equal-proc a b rec-equal?) (hash-proc a rec-hash)))))) (values :gen (generic-instance? :gen))))
-(define-values (equal-proc-accessor) (generics-accessor gen:equal+hash (quote equal-proc)))
 (define-values (hash-proc-accessor) (generics-accessor gen:equal+hash (quote hash-proc)))
 (define-values ($default-record-equal-proc) ($current-record-equal-procedure))
 (define-values ($default-record-hash-proc) ($current-record-hash-procedure))
-(define-values (unsafe-record-type-equal-procedure) (lambda (rtd) (if (gen:equal+hash? rtd) (equal-proc-accessor rtd) $default-record-equal-proc)))
 (define-values (unsafe-record-type-hash-procedure) (lambda (rtd) (if (gen:equal+hash? rtd) (hash-proc-accessor rtd) $default-record-hash-proc)))
-(define-values (record-equal?) (lambda (a b rec-equal?) (letrec-values (((a-rtd) ($record-rtd a)) ((b-rtd) ($record-rtd b))) (if (eq? a-rtd b-rtd) (let-values (((equal-proc) (unsafe-record-type-equal-procedure a-rtd))) (equal-proc a b rec-equal?)) #f))))
 (define-values (record-hash) (lambda (r rec-hash) (letrec-values (((rtd) ($record-rtd r)) ((hash-proc) (unsafe-record-type-hash-procedure rtd))) (hash-proc r rec-hash))))
-(define-values (record-type-equal-procedure) (lambda (rtd) (if (record-type-descriptor? rtd) (void) (let-values () (error (quote record-type-equal-procedure) "expected a record type descriptor" rtd))) (unsafe-record-type-equal-procedure rtd)))
 (define-values (record-type-hash-procedure) (lambda (rtd) (if (record-type-descriptor? rtd) (void) (let-values () (error (quote record-type-hash-procedure) "expected a record type descriptor" rtd))) (unsafe-record-type-hash-procedure rtd)))
 (define-values (gen:custom-write gen:custom-write?) (let-values (((:gen) ($make-generics-type (quote custom-write) (quote ((write-proc v port mode)))))) (values :gen (generic-instance? :gen))))
 (define-values (write-proc-accessor) (generics-accessor gen:custom-write (quote write-proc)))
