+++ src/core/c/list.c
@@ -47,225 +47,6 @@ long improper_list_length(mobj xs) {
     return len;
 }
 
-// Makes an association list.
-// Unsafe: only iterates on `xs`.
-mobj make_assoc(mobj xs, mobj ys) {
-    mobj assoc, it;
-
-    if (minim_nullp(xs))
-        return minim_null;
-
-    assoc = Mcons(Mcons(minim_car(xs), minim_car(ys)), minim_null);
-    it = assoc;
-    while (!minim_nullp(xs = minim_cdr(xs))) {
-        ys = minim_cdr(ys);
-        minim_cdr(it) = Mcons(Mcons(minim_car(xs), minim_car(ys)), minim_null);
-        it = minim_cdr(it);
-    }
-
-    return assoc;
-}
-
-void for_each(mobj proc, int argc, mobj *args, mobj env) {
-    mobj *lsts;
-    long len0, len;
-    int stashc, i;
-
-    if (!minim_procp(proc))
-        bad_type_exn("for-each", "procedure?", proc);
-
-    for (i = 0; i < argc; ++i) {
-        if (!minim_listp(args[i]))
-            bad_type_exn("for-each", "list?", args[i]);
-        
-        if (i == 0) {
-            len0 = list_length(args[i]);
-        } else {
-            len = list_length(args[i]);
-            if (len != len0) {
-                fprintf(stderr, "for-each: lists of different lengths\n");
-                fprintf(stderr, "  one list: %ld, second list: %ld\n", len0, len);
-                minim_shutdown(1);
-            }
-        }
-    }
-
-    // stash lists since we call the procedure
-    // TODO: potential for GC to lose track of the head of each list
-    lsts = GC_alloc(argc * sizeof(mobj));
-    memcpy(lsts, args, argc * sizeof(mobj));
-
-    stashc = stash_call_args();
-    while (!minim_nullp(lsts[0])) {
-        for (i = 0; i < argc; ++i) {
-            push_call_arg(minim_car(lsts[i]));
-            lsts[i] = minim_cdr(lsts[i]);
-        }
-
-        // ignore the result (side-effect only)
-        call_with_args(proc, env);
-    }
-
-    prepare_call_args(stashc);
-}
-
-mobj map_list(mobj proc, int argc, mobj *args, mobj env) {
-    mobj *lsts, res, head, tail;
-    minim_thread *th;
-    long len0, len;
-    int stashc, i;
-
-    if (!minim_procp(proc))
-        bad_type_exn("map", "procedure?", proc);
-
-    for (i = 0; i < argc; ++i) {
-        if (!minim_listp(args[i]))
-            bad_type_exn("map", "list?", args[i]);
-        
-        if (i == 0) {
-            len0 = list_length(args[i]);
-        } else {
-            len = list_length(args[i]);
-            if (len != len0) {
-                fprintf(stderr, "map: lists of different lengths\n");
-                fprintf(stderr, "  one list: %ld, second list: %ld\n", len0, len);
-                minim_shutdown(1);
-            }
-        }
-    }
-
-    // stash lists since we call the procedure
-    // TODO: potential for GC to lose track of the head of each list
-    lsts = GC_alloc(argc * sizeof(mobj));
-    memcpy(lsts, args, argc * sizeof(mobj));
-
-    head = NULL;
-    stashc = stash_call_args();
-    while (!minim_nullp(lsts[0])) {
-        for (i = 0; i < argc; ++i) {
-            push_call_arg(minim_car(lsts[i]));
-            lsts[i] = minim_cdr(lsts[i]);
-        }
-
-        res = call_with_args(proc, env);
-        if (minim_valuesp(res)) {
-            th = current_thread();
-            if (values_buffer_count(th) != 1) {
-                fprintf(stderr, "result arity mismatch\n");
-                fprintf(stderr, "  expected: 1, received: %d\n", values_buffer_count(th));
-                minim_shutdown(1);
-            } else {
-                res = values_buffer_ref(th, 0);
-            }
-        }
-
-        if (head) {
-            minim_cdr(tail) = Mcons(res, minim_null);
-            tail = minim_cdr(tail);
-        } else {
-            head = Mcons(res, minim_null);
-            tail = head;
-        }
-    }
-
-    prepare_call_args(stashc);
-    return (head ? head : minim_null);
-}
-
-mobj andmap(mobj proc, int argc, mobj *args, mobj env) {
-    mobj *lsts;
-    long len0, len;
-    int stashc, i;
-
-    if (!minim_procp(proc))
-        bad_type_exn("andmap", "procedure?", proc);
-
-    for (i = 0; i < argc; ++i) {
-        if (!minim_listp(args[i]))
-            bad_type_exn("andmap", "list?", args[i]);
-        
-        if (i == 0) {
-            len0 = list_length(args[i]);
-        } else {
-            len = list_length(args[i]);
-            if (len != len0) {
-                fprintf(stderr, "andmap: lists of different lengths\n");
-                fprintf(stderr, "  one list: %ld, second list: %ld\n", len0, len);
-                minim_shutdown(1);
-            }
-        }
-    }
-
-    // stash lists since we call the procedure
-    // TODO: potential for GC to lose track of the head of each list
-    lsts = GC_alloc(argc * sizeof(mobj));
-    memcpy(lsts, args, argc * sizeof(mobj));
-
-    stashc = stash_call_args();
-    while (!minim_nullp(lsts[0])) {
-        for (i = 0; i < argc; ++i) {
-            push_call_arg(minim_car(lsts[i]));
-            lsts[i] = minim_cdr(lsts[i]);
-        }
-
-        // only check for false (early exit)
-        if (minim_falsep(call_with_args(proc, env))) {
-            prepare_call_args(stashc);
-            return minim_false;
-        }
-    }
-
-    prepare_call_args(stashc);
-    return minim_true;
-}
-
-mobj ormap(mobj proc, int argc, mobj *args, mobj env) {
-    mobj *lsts;
-    long len0, len;
-    int stashc, i;
-
-    if (!minim_procp(proc))
-        bad_type_exn("ormap", "procedure?", proc);
-
-    for (i = 0; i < argc; ++i) {
-        if (!minim_listp(args[i]))
-            bad_type_exn("ormap", "list?", args[i]);
-        
-        if (i == 0) {
-            len0 = list_length(args[i]);
-        } else {
-            len = list_length(args[i]);
-            if (len != len0) {
-                fprintf(stderr, "ormap: lists of different lengths\n");
-                fprintf(stderr, "  one list: %ld, second list: %ld\n", len0, len);
-                minim_shutdown(1);
-            }
-        }
-    }
-
-    // stash lists since we call the procedure
-    // TODO: potential for GC to lose track of the head of each list
-    lsts = GC_alloc(argc * sizeof(mobj));
-    memcpy(lsts, args, argc * sizeof(mobj));
-
-    stashc = stash_call_args();
-    while (!minim_nullp(lsts[0])) {
-        for (i = 0; i < argc; ++i) {
-            push_call_arg(minim_car(lsts[i]));
-            lsts[i] = minim_cdr(lsts[i]);
-        }
-
-        // only check for false (early exit)
-        if (!minim_falsep(call_with_args(proc, env))) {
-            prepare_call_args(stashc);
-            return minim_true;
-        }
-    }
-
-    prepare_call_args(stashc);
-    return minim_false;
-}
-
 //
 //  Primitives
 //
